Move unit conversion code out of header, and reduce header dependencies
